feat: add ordinals addresses in receive modal
